model.py
import whisper
import random

model = whisper.load_model("base")

# Dummy placeholder function — will refine later
def classify_accent(audio_path: str):
    try:
        result = model.transcribe(audio_path)
        transcript = result["text"]

        # Simulated logic — can replace with actual classifier later
        fake_accents = ["British", "American", "Australian", "Indian"]
        accent = random.choice(fake_accents)
        confidence = round(random.uniform(75, 98), 2)

        explanation = f"The accent was determined based on phoneme patterns and speech rhythm detected in the transcription: \"{transcript[:80]}...\""

        return {
            "accent": accent,
            "confidence": confidence,
            "explanation": explanation
        }
    except Exception as e:
        print(f"Error in classify_accent: {e}")
        return None
