requirements.txt
streamlit
moviepy
openai-whisper
torch
torchaudio
pytube
requests
